refactor: move ChatsListItem to global components
